chore: cleanup final code
