+++ frameworks-plugins/angular-row-detail-plugin/src/angularRowDetailView.ts
@@ -66,8 +66,6 @@ export class AngularRowDetailView extends UniversalSlickRowDetailView {
   }
 
   get rowDetailViewOptions(): RowDetailView | undefined {
-    // Read from getOptions() (which returns _addonOptions) instead of gridOptions.rowDetailView
-    // so that dynamic updates via setOptions() are reflected
     return this.getOptions() as RowDetailView | undefined;
   }
