+++ packages/utils/src/index.ts
@@ -1,5 +1,4 @@
 export * from './domUtils.js';
-export * from './errorHandlingUtil.js';
 export * from './nodeExtend.js';
 export * from './stripTagsUtil.js';
 export type * from './models/interfaces.js';
